+++ q1pulse/lang/timed_statements.py
@@ -6,12 +6,6 @@ class TimedStatement(Statement):
         self.time = time
 
 
-class MultiBranchStatement(TimedStatement):
-    def __init__(self, time):
-        super().__init__(time)
-        self.branches = []
-
-
 class WaitRegStatement(TimedStatement):
     def __init__(self, time, register):
         super().__init__(time)