+++ tests/init_logger.py
@@ -1,7 +1,7 @@
 
 import logging
 
-logging.basicConfig(filename='app.log', filemode='w',
+logging.basicConfig(filename='log.txt', filemode='w',
                     format='%(asctime)s %(name)s - %(levelname)s - %(message)s',
                     level=logging.DEBUG)
 logging.getLogger('matplotlib').setLevel(logging.INFO)