+++ tests/test_event_feedback.py
@@ -3,6 +3,7 @@ import numpy as np
 from q1pulse.instrument import Q1Instrument
 from q1pulse.lang.program_variables import IntVariable
 
+import init_logger
 from init_pulsars import qcm0, qrm1, q1asm_isa_v2
 
 from plot_util import plot_output